Create studentGradeCounter.cpp
This C++ program will display 3 main menu, which is input, grade output, and exit. As the name implies, you can input the student's grades into the first menu (as many as you want), and then get the output at the grade output menu. The statistic data displayed are: total and average of the each of students' grade (UTS, UAS, and Tugas), total of all of the grades, and the average of all of the grades combined.